middleware/requirements.txt
Flask==3.0.3
Flask-Cors==4.0.1
Jinja2==3.1.4
kafka-python==2.0.2
pandas==1.3.1